avoid possible NPE if something stops early
git-svn-id: http://fusesource.com/forge/svn/cloudmix/trunk@213 c32a53b6-b962-0410-9d7f-cad4864ec771

+++ org.fusesource.cloudmix.agent.mop/src/main/java/org/fusesource/cloudmix/agent/mop/MopProcess.java
@@ -94,9 +94,11 @@ public class MopProcess {
     }
 
     public void stop() throws Exception {
-        ProcessRunner processRunner = mop.getProcessRunner();
-        if (processRunner != null) {
-            processRunner.kill();
+        if (mop != null) {
+            ProcessRunner processRunner = mop.getProcessRunner();
+            if (processRunner != null) {
+                processRunner.kill();
+            }
         }
         mop = null;
         thread = null;
